// src/api/api-routes.enum.ts
export enum ApiRoutes {
  Auth = "/auth",
  MssEventTypes = "/MssEventTypes",
  CtrlEventTypes = "/CtrlEventTypes",
  Events = "/events",
  EventsForUser = "/get-events-for-user",
  CtrlEventsForUser = "/pspcontrol/ctrlEvents/get-ctrlevents-for-user",
  EventsWidget = "/get-events-widget",
  AckStates = "/AckStates",
  SiknRsus = "/siknrsus",
  ConstantRisks = "/constantrisks",
  BindConstantRisks = "/siknrsus/bind",
  MssEventSeverityLevels = "/MssEventSeverityLevels",
  CreateConstantRisk = "/constantrisks",
  EventsCharts = "/get-events-charts",
  CheckingObjects = "/pspcontrol/checkingobjects",
  GetFilter = "/filterDescription",
  GetFilteredValues = "/filterValues",
  GetVerificationSchedulesList = "/pspcontrol/VerificationSchedules/list",
  VerificationSchedules = "/pspcontrol/VerificationSchedules",
  GetVerificationLevels = "/pspcontrol/VerificationLevels",
  GetPsp = "/PspControl/pspcard",
  VerificationActs = "/pspcontrol/act",
  GetActionPlans = "/pspcontrol/plan/filter",
  GetActionPlansFilterDescription = "/pspcontrol/plan/filterDescription",
  GetActionPlansFilterValues = "/pspcontrol/plan/filterValues",
  GetAlgorithmTree = "/algs/get-status",
  GetAlgConfiguration = "/algs/get-alg-configuration",
  GetAlgOperands = "/algs/get-alg-operands",
  GetAlgTemplates = "/algs/get-alg-templates",
  GetOperands = "/algs/get-operands",
  GetConfiguration = "/algs/get-configuration",
  GetActionPlan = "/pspcontrol/act",
  Plan = "/pspcontrol/plan",
  FAQ = "/faq",
  operandsAlgorithms = "/algs/get-operand-algs?elementId=",
  operandsAlgorithmsAf = "/algs/get-operand-afAlgs?elementId=",
  PostSetPoints = "/algs/update-setpoints",
  TypicalPlan = "/pspcontrol/typicalPlan",
  Elimination = "/pspcontrol/elimination",
  Acquaintance = "/pspcontrol/acquaintance",
  ProgramKspp = "/pspcontrol/programKspp",
  EliminationTypical = "/pspcontrol/eliminationTypical",
  Dictionaries = "/pspcontrol/dictionaries",
  CtrlEvents = "/pspcontrol/CtrlEvents",
  Notification = "/notification",
  BdmiLastDate = "/pspcontrol/bdmilastdate",
}